Test url cleanup - Joining project/workspace  pytest executor into one and moving it to the pytest folder - adding specific ContainerDiscoverer uri strings for project and workspace - adding specific executor uri strings for pytest - removing TestMethodResolver.cs. Test Explorer will just run all tests in file with right clicking in editor and selecting "Run Test".
