+++ Python/Product/TestAdapter.Executor/Pytest/TestExecutorPytest.cs
@@ -1,0 +1,171 @@
+// Python Tools for Visual Studio
+// Copyright(c) Microsoft Corporation
+// All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the License); you may not use
+// this file except in compliance with the License. You may obtain a copy of the
+// License at http://www.apache.org/licenses/LICENSE-2.0
+//
+// THIS CODE IS PROVIDED ON AN  *AS IS* BASIS, WITHOUT WARRANTIES OR CONDITIONS
+// OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING WITHOUT LIMITATION ANY
+// IMPLIED WARRANTIES OR CONDITIONS OF TITLE, FITNESS FOR A PARTICULAR PURPOSE,
+// MERCHANTABILITY OR NON-INFRINGEMENT.
+//
+// See the Apache Version 2.0 License for specific language governing
+// permissions and limitations under the License.
+
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+using System.IO;
+using System.Linq;
+using System.Net;
+using System.Net.Sockets;
+using System.Runtime.InteropServices;
+using System.Security.Cryptography;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+using System.Xml.XPath;
+using Microsoft.PythonTools.Analysis;
+using Microsoft.PythonTools.Infrastructure;
+using Microsoft.PythonTools.Ipc.Json;
+using Microsoft.PythonTools.TestAdapter.Config;
+using Microsoft.PythonTools.TestAdapter.Pytest;
+using Microsoft.PythonTools.TestAdapter.Services;
+using Microsoft.VisualStudio.TestPlatform.ObjectModel;
+using Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter;
+using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
+using TP = Microsoft.PythonTools.TestAdapter.TestProtocol;
+
+namespace Microsoft.PythonTools.TestAdapter {
+
+    [ExtensionUri(PythonConstants.PytestExecutorUriString)]
+
+    [SuppressMessage("Microsoft.Design", "CA1001:TypesThatOwnDisposableFieldsShouldBeDisposable",
+        Justification = "object owned by VS")]
+    public class TestExecutorPytest : ITestExecutor {
+        private static readonly Guid PythonRemoteDebugPortSupplierUnsecuredId = new Guid("{FEB76325-D127-4E02-B59D-B16D93D46CF5}");
+        private static readonly Guid PythonDebugEngineGuid = new Guid("EC1375B7-E2CE-43E8-BF75-DC638DE1F1F9");
+        private static readonly Guid NativeDebugEngineGuid = new Guid("3B476D35-A401-11D2-AAD4-00C04F990171");
+
+        private static readonly string TestLauncherPath = PythonToolsInstallPath.GetFile("visualstudio_py_testlauncher.py");
+        internal static readonly Uri PythonCodeCoverageUri = new Uri("datacollector://Microsoft/PythonCodeCoverage/1.0");
+
+        private readonly ManualResetEvent _cancelRequested = new ManualResetEvent(false);
+
+        private readonly VisualStudioProxy _app;
+
+        public TestExecutorPytest() {
+            _app = VisualStudioProxy.FromEnvironmentVariable(PythonConstants.PythonToolsProcessIdEnvironmentVariable);
+        }
+
+        public void Cancel() {
+            _cancelRequested.Set();
+        }
+
+        public void RunTests(IEnumerable<string> sources, IRunContext runContext, IFrameworkHandle frameworkHandle) {
+
+            //MessageBox.Show("Hello1: " + Process.GetCurrentProcess().Id);
+
+            if (sources == null) {
+                throw new ArgumentNullException(nameof(sources));
+            }
+
+            if (runContext == null) {
+                throw new ArgumentNullException(nameof(runContext));
+            }
+
+            if (frameworkHandle == null) {
+                throw new ArgumentNullException(nameof(frameworkHandle));
+            }
+
+            _cancelRequested.Reset();
+
+            var sourceToProjSettings = RunSettingsUtil.GetSourceToProjSettings(runContext.RunSettings);
+
+            var testColletion = new TestCollection();
+
+            foreach (var testGroup in sources.GroupBy(x => sourceToProjSettings[x])) {
+                var settings = testGroup.Key;
+
+                try {
+                    var discovery = DiscovererFactory.GetDiscoverer(settings);
+                    discovery.DiscoverTests(testGroup, frameworkHandle, testColletion);
+                } catch (Exception ex) {
+                    frameworkHandle.SendMessage(TestMessageLevel.Error, ex.Message);
+                }
+
+                if (_cancelRequested.WaitOne(0)) {
+                    return;
+                }
+            }
+
+            RunPytest(testColletion.Tests, runContext, frameworkHandle);
+        }
+
+        public void RunTests(IEnumerable<TestCase> tests, IRunContext runContext, IFrameworkHandle frameworkHandle) {
+
+            //MessageBox.Show("Hello1: " + Process.GetCurrentProcess().Id);
+
+            if (tests == null) {
+                throw new ArgumentNullException(nameof(tests));
+            }
+
+            if (runContext == null) {
+                throw new ArgumentNullException(nameof(runContext));
+            }
+
+            if (frameworkHandle == null) {
+                throw new ArgumentNullException(nameof(frameworkHandle));
+            }
+
+            RunPytest(tests, runContext, frameworkHandle);
+
+            _cancelRequested.Reset();
+        }
+
+        private void RunPytest(
+            IEnumerable<TestCase> tests,
+            IRunContext runContext,
+            IFrameworkHandle frameworkHandle
+        ) {
+            var sourceToProjSettings = RunSettingsUtil.GetSourceToProjSettings(runContext.RunSettings);
+
+            foreach (var testGroup in tests.GroupBy(t => sourceToProjSettings.TryGetValue(t.CodeFilePath, out PythonProjectSettings proj) ? proj : null)) {
+                if (testGroup.Key == null) {
+                    Debug.WriteLine("Missing projectSettings for TestCases:");
+                    Debug.WriteLine(String.Join(",\n", testGroup));
+                }
+
+                if (_cancelRequested.WaitOne(0)) {
+                    break;
+                }
+
+                RunTestGroup(testGroup, runContext, frameworkHandle);
+            }
+        }
+
+        private void RunTestGroup(IGrouping<PythonProjectSettings, TestCase> testGroup, IRunContext runContext, IFrameworkHandle frameworkHandle) {
+            PythonProjectSettings settings = testGroup.Key;
+            if (settings.TestFramwork != TestFrameworkType.Pytest) {
+                return;
+            }
+
+            using (var executor = new ExecutorService(settings, frameworkHandle, runContext)) {
+                var resultsXML = executor.Run(testGroup);
+
+                var testResults = TestResultParser.Parse(resultsXML, testGroup);
+                foreach (var result in testResults) {
+
+                    if (_cancelRequested.WaitOne(0)) {
+                        break;
+                    }
+                    frameworkHandle.RecordResult(result);
+                }
+            }
+        }
+    }
+}
